pyproject.toml
[tool.poetry]
name = "alicerce"
version = "0.1.0"
description = ""
authors = ["Jessica Sousa <jessicacardosops@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^1.1.2"
flask-restx = "^0.2.0"
Flask-SQLAlchemy = "^2.4.4"
Flask-Bcrypt = "^0.7.1"
dynaconf = "^3.1.2"
Flask-Migrate = "^2.5.3"
Flask-JWT-Extended = "^3.25.0"
flask_simplelogin = "^0.0.7"
Flask-Admin = "^1.5.7"
Bootstrap-Flask = "^1.5.1"
python-dotenv = "^0.15.0"
Flask-Limiter = "^1.4"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
rope = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
